[rules] moved rules folder up level to be accessed by both ui and api
