+++ rules/movement/movement-validate-path.js
