+++ ui/client/src/components/Board.jsx
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import Square from './Square.jsx';
 import axios from 'axios';
-import { rotateBoard } from '../../rules/utilities/';
+import { rotateBoard } from '../../../../rules/utilities/';
 
 axios.defaults.withCredentials = true;
