+++ ui/client/src/components/MoveHistory.jsx
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 import Result from './Result.jsx';
-import { printMoves } from '../../rules/utilities/';
+import { printMoves } from '../../../../rules/utilities/';
 
 class MoveHistory extends Component {
   constructor(props) {
