+++ ui/client/src/components/PlayerCard.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import axios from 'axios';
-import { printCapturedPieces } from '../../rules/utilities/';
+import { printCapturedPieces } from '../../../../rules/utilities/';
 
 axios.defaults.withCredentials = true;
